+++ src/components/styles/common.scss
@@ -2,6 +2,7 @@
   @import url('https://fonts.googleapis.com/css2?family=DM+Sans:ital,wght@0,400;0,500;0,700;1,400;1,500;1,700&family=Raleway:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800&family=Red+Rose:wght@300;400;500;600;700&display=swap');
 
 $clear-pink: #F5ECE9;
+$pink: #EADDD7;
 $chocolate: #30221E;
 $title-font: 'Red Rose';
 $text-font: 'Raleway';
